fix tablo pumse

+++ database/migrations/2022_10_13_073209_create_grades_table.php
@@ -15,7 +15,7 @@ class CreateGradesTable extends Migration
     {
         Schema::create('grades', function (Blueprint $table) {
             $table->id();
-            $table->integer('grade')->nullable();
+            $table->float('grade')->nullable();
             $table->integer('competitor_id')->nullable();
             $table->timestamps();
         });
